A5-Advanced REST Features/Marina-API/requirements.txt
Flask==2.2.2
google-cloud-datastore==2.8.2
Werkzeug==2.3.7